+++ src/lib/lesson-metadata/__tests__/git-metadata.test.ts
@@ -56,8 +56,8 @@ import {
  */
 const PAGES_ROOT = "src/pages";
 const FALLBACK_AUTHOR = "Proyecto DIBS";
-const SAMPLE_SOURCE = "src/pages/notes/software-libraries/scripting/first-script/index.astro";
-const SAMPLE_ROUTE = "/notes/software-libraries/scripting/first-script/";
+const SAMPLE_SOURCE = "src/pages/notes/scripting/first-script/index.astro";
+const SAMPLE_ROUTE = "/notes/scripting/first-script/";
 
 describe("sourceFileToLessonPath", () => {
     /**
@@ -75,7 +75,7 @@ describe("sourceFileToLessonPath", () => {
         ["src/pages/foo/bar.astro", "/foo/bar/"],
         ["src/pages//foo//index.astro", "/foo/"],
         [
-            "src\\pages\\notes\\software-libraries\\scripting\\first-script\\index.astro",
+            "src\\pages\\notes\\scripting\\first-script\\index.astro",
             SAMPLE_ROUTE,
         ],
     ])("maps %s to %s", (input, expected) => {